Use `runsolver -v` to fetch the basic run stats
`runsolver -v` stores the most relevant run information
(namely timeout status, memout status, and CPU time)
in an easy-to-parse format.
This commit ensures that `AbstractWrapper`
invokes `runsolver` with a `-v` output file specified
and that the output file is parsed
to determine the timeout status, memout status, and CPU time
of the sandboxed command.